+++ pkg/codegen/schema/schema.go
@@ -18,6 +18,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"math"
+	"net/url"
 	"regexp"
 	"sort"
 	"strings"
@@ -49,6 +50,7 @@ const (
 	archiveType primitiveType = 5
 	assetType   primitiveType = 6
 	anyType     primitiveType = 7
+	jsonType    primitiveType = 8
 )
 
 func (t primitiveType) String() string {
@@ -65,6 +67,8 @@ func (t primitiveType) String() string {
 		return "pulumi:pulumi:Archive"
 	case assetType:
 		return "pulumi:pulumi:Asset"
+	case jsonType:
+		fallthrough
 	case anyType:
 		return "pulumi:pulumi:Any"
 	default:
@@ -94,6 +98,8 @@ var (
 	ArchiveType Type = archiveType
 	// AssetType represents the set of Pulumi Asset values.
 	AssetType Type = assetType
+	// JSONType represents the set of JSON-encoded values.
+	JSONType Type = jsonType
 	// AnyType represents the complete set of values.
 	AnyType Type = anyType
 )
@@ -793,6 +799,8 @@ func (t *types) bindType(spec TypeSpec) (Type, error) {
 			return ArchiveType, nil
 		case "pulumi.json#/Asset":
 			return AssetType, nil
+		case "pulumi.json#/Json":
+			return JSONType, nil
 		case "pulumi.json#/Any":
 			return AnyType, nil
 		}
@@ -802,7 +810,10 @@ func (t *types) bindType(spec TypeSpec) (Type, error) {
 			return nil, errors.Errorf("failed to parse ref %s", spec.Ref)
 		}
 
-		token := spec.Ref[len("#/types/"):]
+		token, err := url.PathUnescape(spec.Ref[len("#/types/"):])
+		if err != nil {
+			return nil, errors.Errorf("failed to parse ref %s", spec.Ref)
+		}
 		if typ, ok := t.objects[token]; ok {
 			return typ, nil
 		}