+++ CHANGELOG.md
@@ -2,7 +2,12 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-_(none)_
+
+- Add support for untagged outputs in Go SDK.
+ [#4640](https://github.com/pulumi/pulumi/pull/4640)
+
+- Fix a Regression in .NET unit testing.
+ [#4656](https://github.com/pulumi/pulumi/pull/4656)
 
 ## 2.2.1 (2020-05-13)
 - Add new brew target to fix homebrew builds